+++ packages/components/src/index.ts
@@ -24,7 +24,7 @@ export { Avatar }     from './components/Avatar'
 export type { AvatarProps, AvatarSize, AvatarStatus } from './components/Avatar'
 
 export { Toast }      from './components/Toast'
-export type { ToastProps, ToastTone, ToastVariant } from './components/Toast'
+export type { ToastProps, ToastTone, ToastVariant, ToastAppearance } from './components/Toast'
 
 export { Listbox } from './components/Listbox'
 export type { ListboxProps, ListboxOption, ListboxSize } from './components/Listbox'